refactor: 타임스탬프 함수 재활용
util파일에서 파베 타임스탬스 날짜 변경 함수 관리합니다.

by @harenohee
